requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
langfuse>=2.60.0,<3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
joblib>=1.4.0

